requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
importlib-metadata==2.0.0
mccabe==0.6.1
psycopg2==2.7.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
uuid==1.30
whitenoise==5.2.0
zipp==3.4.0
